+++ Lagrange.Core/Internal/Events/Message/GetC2CMessageEvent.cs
@@ -2,13 +2,13 @@ using Lagrange.Core.Internal.Packets.Message;
 
 namespace Lagrange.Core.Internal.Events.Message;
 
-internal class GetC2CMessageEventReq(string peerUid, int startSequence, int endSequence) : ProtocolEvent
+internal class GetC2CMessageEventReq(string peerUid, ulong startSequence, ulong endSequence) : ProtocolEvent
 {
     public string PeerUid { get; } = peerUid;
 
-    public int StartSequence { get; } = startSequence;
+    public ulong StartSequence { get; } = startSequence;
 
-    public int EndSequence { get; } = endSequence;
+    public ulong EndSequence { get; } = endSequence;
 }
 
 internal class GetC2CMessageEventResp(List<CommonMessage> chains) : ProtocolEvent
