+++ Lagrange.Core/Internal/Packets/Message/NTSendMessage.cs
@@ -13,7 +13,7 @@ internal partial class PbSendMsgReq
     
     [ProtoMember(3)] public MessageBody MessageBody { get; set; }
     
-    [ProtoMember(4)] public int ClientSequence { get; set; }
+    [ProtoMember(4)] public ulong ClientSequence { get; set; }
     
     [ProtoMember(5)] public uint Random { get; set; }
 }
